+++ packages/core-js/es/symbol/index.js
@@ -2,6 +2,7 @@ require('../../modules/es.array.concat');
 require('../../modules/es.object.to-string');
 require('../../modules/es.symbol');
 require('../../modules/es.symbol.async-iterator');
+require('../../modules/es.symbol.description');
 require('../../modules/es.symbol.has-instance');
 require('../../modules/es.symbol.is-concat-spreadable');
 require('../../modules/es.symbol.iterator');