+++ packages/core-js/features/symbol/index.js
@@ -1,6 +1,5 @@
 module.exports = require('../../es/symbol');
 
-require('../../modules/esnext.symbol.description');
 require('../../modules/esnext.symbol.dispose');
 require('../../modules/esnext.symbol.observable');
 require('../../modules/esnext.symbol.pattern-match');