+++ packages/core-js/stage/3.js
@@ -3,6 +3,5 @@ require('../proposals/global-this');
 require('../proposals/object-from-entries');
 require('../proposals/string-match-all');
 require('../proposals/string-trim-start-end');
-require('../proposals/symbol-description');
 
 module.exports = require('./4');